separated device classes to independent package
