--- simulator/SicTools/src/sicxe/sim/vm/Device.java
@@ -1,24 +1,0 @@
-package sicxe.sim.vm;
-
-/**
- * Base class for SIC/XE device.
- * @author jure
- */
-public class Device {
-
-    public boolean test() {
-        // always true: EOF is returned by read
-        return true;
-    }
-
-    public int read() {
-        return -1;
-    }
-
-    public void write(int value) {
-    }
-
-    public void reset() {
-    }
-
-}
